- fixed the usedcolor array's base type.
The usedcolors array which counts the number of pixels in a given color in a font used bytes as storage, so any color that just happened to have a count that is a multiple of 256 the color was considered not present.

+++ src/gamedata/fonts/fontinternals.h
@@ -39,6 +39,6 @@ extern uint16_t upperforlower[65536];
 
 class FImageSource;
 
-void RecordTextureColors (FImageSource *pic, uint8_t *usedcolors);
+void RecordTextureColors (FImageSource *pic, uint32_t *usedcolors);
 bool myislower(int code);
 int stripaccent(int code);
